+++ packages/stream_feeds/lib/src/client/feeds_client_impl.dart
@@ -199,7 +199,7 @@ class StreamFeedsClientImpl implements StreamFeedsClient {
     environment: const SystemEnvironment(
       sdkName: 'stream-feeds-dart',
       sdkIdentifier: 'dart',
-      sdkVersion: '0.1.0',
+      sdkVersion: '0.3.0',
     ),
   );